using nextjs image optimization for uploads and 404 page design
